renamed the DAL directory (Data Access Layer) to Data
